+++ src/test/kotlin/org/opensearch/indexmanagement/snapshotmanagement/engine/states/creation/CreationStartStateTests.kt
@@ -6,21 +6,21 @@
 package org.opensearch.indexmanagement.snapshotmanagement.engine.states.creation
 
 import kotlinx.coroutines.runBlocking
+import org.opensearch.indexmanagement.MocksTestCase
 import org.opensearch.indexmanagement.snapshotmanagement.engine.SMStateMachine
-import org.opensearch.indexmanagement.snapshotmanagement.engine.SMStateMachineTests
 import org.opensearch.indexmanagement.snapshotmanagement.engine.states.SMResult
 import org.opensearch.indexmanagement.snapshotmanagement.engine.states.SMState
 import org.opensearch.indexmanagement.snapshotmanagement.randomSMMetadata
 import org.opensearch.indexmanagement.snapshotmanagement.randomSMPolicy
 
-class CreationStartStateTests : SMStateMachineTests() {
+class CreationStartStateTests : MocksTestCase() {
 
     fun `test start state execution`() = runBlocking {
         val metadata = randomSMMetadata(
             creationCurrentState = SMState.CREATION_FINISHED
         )
         val job = randomSMPolicy()
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.CREATION_START.instance.execute(context)
         assertTrue("Execution result should be Next.", result is SMResult.Next)
