+++ src/test/kotlin/org/opensearch/indexmanagement/snapshotmanagement/engine/states/deletion/DeletingStateTests.kt
@@ -8,8 +8,8 @@ package org.opensearch.indexmanagement.snapshotmanagement.engine.states.deletion
 import kotlinx.coroutines.runBlocking
 import org.opensearch.action.support.master.AcknowledgedResponse
 import org.opensearch.common.unit.TimeValue
+import org.opensearch.indexmanagement.MocksTestCase
 import org.opensearch.indexmanagement.snapshotmanagement.engine.SMStateMachine
-import org.opensearch.indexmanagement.snapshotmanagement.engine.SMStateMachineTests
 import org.opensearch.indexmanagement.snapshotmanagement.engine.states.SMResult
 import org.opensearch.indexmanagement.snapshotmanagement.engine.states.SMState
 import org.opensearch.indexmanagement.snapshotmanagement.mockGetSnapshotResponse
@@ -22,7 +22,7 @@ import org.opensearch.indexmanagement.snapshotmanagement.randomSMPolicy
 import org.opensearch.indexmanagement.snapshotmanagement.randomSnapshotName
 import java.time.Instant.now
 
-class DeletingStateTests : SMStateMachineTests() {
+class DeletingStateTests : MocksTestCase() {
 
     fun `test snapshots exceed max count`() = runBlocking {
         mockGetSnapshotsCall(response = mockGetSnapshotResponse(11))
@@ -35,7 +35,7 @@ class DeletingStateTests : SMStateMachineTests() {
             policyName = "daily-snapshot",
             deletionMaxCount = 10,
         )
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Next.", result is SMResult.Next)
@@ -60,7 +60,7 @@ class DeletingStateTests : SMStateMachineTests() {
             deletionMaxAge = TimeValue.timeValueMinutes(1),
             deletionMinCount = 2,
         )
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Next.", result is SMResult.Next)
@@ -85,7 +85,7 @@ class DeletingStateTests : SMStateMachineTests() {
             deletionMaxAge = TimeValue.timeValueMinutes(1),
             deletionMinCount = 2,
         )
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Next.", result is SMResult.Next)
@@ -107,7 +107,7 @@ class DeletingStateTests : SMStateMachineTests() {
             deletionMaxAge = TimeValue.timeValueMinutes(1),
             deletionMinCount = 2,
         )
-        val context = SMStateMachine(client, job, metadata)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Next.", result is SMResult.Next)
@@ -126,7 +126,7 @@ class DeletingStateTests : SMStateMachineTests() {
             deletionCurrentState = SMState.DELETION_CONDITION_MET,
         )
         val job = randomSMPolicy(policyName = "daily-snapshot")
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Failure.", result is SMResult.Fail)
@@ -148,7 +148,7 @@ class DeletingStateTests : SMStateMachineTests() {
             )
         )
         val job = randomSMPolicy(policyName = "daily-snapshot")
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Fail.", result is SMResult.Fail)
@@ -167,7 +167,7 @@ class DeletingStateTests : SMStateMachineTests() {
         val job = randomSMPolicy(
             deletionNull = true
         )
-        val context = SMStateMachine(client, job, metadata, settings, threadPool)
+        val context = SMStateMachine(client, job, metadata, settings, threadPool, indicesManager)
 
         val result = SMState.DELETING.instance.execute(context)
         assertTrue("Execution result should be Fail.", result is SMResult.Fail)
